dictation software with AI built-in/requirements.txt
pytest
cryptography
PyQt6
pynetdicom
pydicom
requests
hidapi
pywinusb
pydirectinput
loguru
sounddevice
numpy
pytest-qt
